refactor: reorganize AWS modules and update version to 2.0.0-dev
