+++ src/cmd.rs
@@ -1,5 +1,2 @@
-pub mod aws_detect;
-pub mod aws_detect_writer;
-pub mod aws_metrics;
-pub mod aws_summary;
+pub mod aws;
 pub mod update;
