+++ src/cmd/aws.rs
@@ -1,0 +1,3 @@
+pub mod aws_detect;
+pub mod aws_metrics;
+pub mod aws_summary;
