+++ src/core.rs
@@ -1,4 +1,7 @@
 pub mod color;
 pub mod rules;
 pub mod scan;
+pub mod summary;
+pub mod timeline;
+pub mod timeline_writer;
 pub mod util;
