+++ src/main.rs
@@ -1,8 +1,8 @@
+use crate::cmd::aws::aws_detect::aws_detect;
+use crate::cmd::aws::aws_metrics::aws_metrics;
+use crate::cmd::aws::aws_summary::aws_summary;
 use chrono::Local;
 use clap::{CommandFactory, Parser};
-use cmd::aws_detect::aws_detect;
-use cmd::aws_metrics::aws_metrics;
-use cmd::aws_summary::aws_summary;
 use cmd::update::start_update_rules;
 use core::color::SuzakuColor::Green;
 use core::util::{check_path_exists, p, set_rayon_threat_number};
